+++ src/musicscan/data/cd.py
@@ -28,7 +28,7 @@ Data objects for music.
 
 # pylint: disable=too-many-instance-attributes
 
-from musicscan.data.flags import Flags, FlagCodes
+from musicscan.data.flags import AlbumFlags, TrackFlagCodes
 from musicscan.data.titletools import ShortTitleIndex, ShortTitle
 from musicscan.data.track import Track
 from musicscan.data.stringtools import build_complete_filename
@@ -46,7 +46,7 @@ class Album():
         self.title_index = ShortTitleIndex()
         self.discs = {}
         self._first_disc = None
-        self.flags = Flags()
+        self.flags = AlbumFlags()
         self.track_count = 0
 
     def import_tag(self, in_tag):
@@ -165,7 +165,7 @@ class Disc():
         for trk in sort_t:
             if not trk.title:
                 trk.title = 'PLACEHOLDER TITLE'
-                trk.flags.add_flag(FlagCodes.m_title)
+                trk.flags.add_flag(TrackFlagCodes.m_title)
             trk.short_title = ShortTitle(trk.title, self.album.title_index)
 
     def report(self):